refactor report for GHA
