+++ plugin.js
@@ -1,38 +1,12 @@
 // @ts-check
 require('console.table')
 const { getNycReportFilename } = require('./task-utils')
-const { existsSync, readFileSync } = require('fs')
+const { existsSync } = require('fs')
 const NYC = require('nyc')
 const debug = require('debug')('code-coverage')
-const ghCore = require('@actions/core')
-const path = require('path')
+const { reportCodeCoverageGHA } = require('./src/utils')
 
 const nycFilename = getNycReportFilename(process.cwd())
-
-function pickCoverageEmoji(percentage) {
-  if (percentage >= 95) {
-    return '✅'
-  }
-  if (percentage >= 90) {
-    return '🏆'
-  }
-  if (percentage >= 80) {
-    return '🥇'
-  }
-  if (percentage >= 70) {
-    return '🥈'
-  }
-  if (percentage >= 60) {
-    return '🥉'
-  }
-  if (percentage >= 50) {
-    return '📈'
-  }
-  if (percentage >= 40) {
-    return '⚠️'
-  }
-  return '🪫'
-}
 
 function registerCodeCoveragePlugin(on, config) {
   require('./task')(on, config)
@@ -76,55 +50,7 @@ function registerCodeCoveragePlugin(on, config) {
 
     if (process.env.GITHUB_ACTIONS) {
       debug('will report code coverage on GitHub Actions')
-      on('after:run', () => {
-        const summaryFilename = path.join('coverage', 'coverage-summary.json')
-        if (!existsSync(summaryFilename)) {
-          debug('cannot find summary file %s', summaryFilename)
-        } else {
-          const summary = JSON.parse(readFileSync(summaryFilename, 'utf8'))
-          if (summary.total) {
-            debug('code coverage summary totals %o', summary.total)
-            const s = summary.total.statements
-            const b = summary.total.branches
-            const f = summary.total.functions
-            const l = summary.total.lines
-            const row = [
-              String(s.pct),
-              `${s.covered}/${s.total}`,
-              String(b.pct),
-              `${b.covered}/${b.total}`,
-              String(f.pct),
-              `${f.covered}/${f.total}`,
-              String(l.pct),
-              `${l.covered}/${l.total}`,
-            ]
-            debug(row)
-
-            ghCore.summary
-              .addHeading('Code coverage')
-              .addTable([
-                [
-                  { data: 'Statements %', header: true },
-                  { data: pickCoverageEmoji(s.pct), header: true },
-                  { data: 'Branches %', header: true },
-                  { data: pickCoverageEmoji(b.pct), header: true },
-                  { data: 'Functions %', header: true },
-                  { data: pickCoverageEmoji(f.pct), header: true },
-                  { data: 'Lines %', header: true },
-                  { data: pickCoverageEmoji(l.pct), header: true },
-                ],
-                row,
-              ])
-              .addLink(
-                '@bahmutov/cypress-code-coverage',
-                'https://github.com/bahmutov/cypress-code-coverage',
-              )
-              .write()
-          } else {
-            debug('could not find totals in %s', summaryFilename)
-          }
-        }
-      })
+      on('after:run', reportCodeCoverageGHA)
     }
   }
